% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2themes.R
\docType{data}
\name{cbPalette}
\alias{cbPalette}
\title{Color-blind friendly palette
see http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/}
\format{An object of class \code{character} of length 8.}
\usage{
cbPalette
}
\description{
Color-blind friendly palette
see http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/
}
\keyword{datasets}
