% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2themes.R
\name{theme_rasilab}
\alias{theme_rasilab}
\title{Load Rasilab ggplot2 theme}
\usage{
theme_rasilab()
}
\value{
old theme
}
\description{
This function loads the ggplot2 theme_classic with Helvetica as base font,
all font sizes set to size 8, axis line width at 0.25pt, and all axis text
and lines in set to black.
}
